+++ src/jdos/ints/Int10_vptable.java
@@ -499,7 +499,7 @@ public class Int10_vptable {
     };
 
 
-    static public /*Bit16u*/int INT10_SetupVideoParameterTable(/*PhysPt*/long basepos) {
+    static public /*Bit16u*/int INT10_SetupVideoParameterTable(/*PhysPt*/int basepos) {
         if (Dosbox.IS_VGA_ARCH()) {
             for (/*Bitu*/int i=0;i<0x40*0x1d;i++) {
                 Memory.phys_writeb((int)(basepos+i),video_parameter_table_vga[i]);
